+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,11 @@
 
 ## Next Release
 
+## 0.2.1-beta.3
+
+- \[router\] Create queues for transfer creation races
+- \[engine, types\] Make timeout parameter optional for setup, conditional transfer, and withdraw
+
 ## 0.2.1-beta.2
 
 - \[engine\] Remove time validation on engine